Split up wer.py
