+++ meeteval/wer/wer/mimo.py
@@ -1,0 +1,87 @@
+@dataclass(frozen=True)
+class MimoErrorRate(ErrorRate):
+    """
+    >>> MimoErrorRate(0, 10, 0, 0, 0, [(0, 0)])
+    MimoErrorRate(errors=0, length=10, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[(0, 0)])
+    >>> MimoErrorRate(0, 10, 0, 0, 0, [(0, 0)]) + MimoErrorRate(10, 10, 0, 0, 10, [(0, 0)])
+    ErrorRate(errors=10, length=20, insertions=0, deletions=0, substitutions=10, error_rate=0.5)
+    """
+    assignment: Tuple[int, ...]
+
+
+def mimo_word_error_rate(
+        reference: 'List[List[str]] | Dict[List[str]]',
+        hypothesis: 'List[str] | Dict[str]',
+) -> MimoErrorRate:
+    """
+    The Multiple Input speaker, Multiple Output channel (MIMO) WER.
+
+    >>> mimo_word_error_rate([['a b c d e f']], ['a b c d e f'])
+    MimoErrorRate(errors=0, length=6, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[(0, 0)])
+
+    # All correct, utterance order between speakers can change
+    >>> mimo_word_error_rate([['a b', 'c d'], ['e f']], ['a b', 'e f c d'])
+    MimoErrorRate(errors=0, length=6, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[(1, 1), (0, 0), (0, 1)])
+    >>> mimo_word_error_rate([['a b', 'c d'], ['e f']], ['a b', 'c d e f'])
+    MimoErrorRate(errors=0, length=6, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[(0, 0), (0, 1), (1, 1)])
+    >>> mimo_word_error_rate([['a b', 'c d'], ['e f']], ['c d', 'a b e f'])
+    MimoErrorRate(errors=0, length=6, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[(0, 1), (1, 1), (0, 0)])
+
+    >>> mimo_word_error_rate({'A': ['a b', 'c d'], 'B': ['e f']},
+    ...                      {'O1': 'c d', 'O2': 'a b e f'})
+    MimoErrorRate(errors=0, length=6, insertions=0, deletions=0, substitutions=0, error_rate=0.0, assignment=[('A', 'O2'), ('B', 'O2'), ('A', 'O1')])
+
+    """
+    from meeteval.wer.matching.mimo_matching import mimo_matching_v3
+
+    def to_list(obj):
+        if isinstance(obj, dict):
+            return obj.values()
+        else:
+            return obj
+
+    def get_keys(obj):
+        if isinstance(obj, dict):
+            return list(obj.keys())
+        else:
+            return list(range(len(obj)))
+
+    reference_keys = get_keys(reference)
+    hypothesis_keys = get_keys(hypothesis)
+
+    reference_values = [to_list(r) for r in to_list(reference)]
+    hypothesis_values = to_list(hypothesis)
+
+    reference_words = [
+        [u.split() for u in ref_chn]
+        for ref_chn in reference_values
+    ]
+    hypothesis_words = [h.split() for h in hypothesis_values]
+    distance, assignment = mimo_matching_v3(reference_words, hypothesis_words)
+
+    assignment = [
+        (reference_keys[r], hypothesis_keys[h]) for r, h in assignment]
+
+    from meeteval.wer.assignment import apply_mimo_assignment
+    reference_new, hypothesis = apply_mimo_assignment(
+        assignment,
+        reference=reference,
+        hypothesis=hypothesis,
+    )
+
+    er = sum([
+        siso_word_error_rate(
+            ' '.join(r),  # Concatenate all utterances from one speaker
+            h,
+        )
+        for r, h in zip(to_list(reference_new), to_list(hypothesis))
+    ])
+    assert er.errors == distance, (distance, er)
+
+    return MimoErrorRate(
+        er.errors, er.length,
+        insertions=er.insertions,
+        deletions=er.deletions,
+        substitutions=er.substitutions,
+        assignment=assignment,
+    )
